1. Add an independent test class (AetgCoveringArrayEngineTest) for AetgCoveringArrayEngine    and remove a test method in TestSuiteBuilderTest. 2. Fix infinite loop that I introduced by (probably) the last commit.
Right now, AetgCoveringArrayEngine generates a following test suite for 3^2 factor space.

{factor1=1, factor2=1, factor3=1}
{factor1=2, factor2=1, factor3=2}
{factor1=3, factor2=1, factor3=3}
{factor1=2, factor2=2, factor3=1}
{factor1=2, factor2=3, factor3=3}
{factor1=3, factor2=2, factor3=1}
{factor1=3, factor2=3, factor3=1}
{factor1=1, factor2=1, factor3=2}
{factor1=1, factor2=1, factor3=3}
{factor1=2, factor2=2, factor3=2}
{factor1=3, factor2=3, factor3=2}
{factor1=2, factor2=2, factor3=3}

This contains 12 test cases and does not cover following pairs

[{factor1=1, factor2=2}, {factor1=1, factor2=3}]

Both should be fixed.
